refactor(data-fetcher): move sql specific utils to their own file
